+++ server/C/cem/sedtrans.h
@@ -9,10 +9,10 @@ extern "C" {
 #include "BeachNode.h"
 #include "BeachGrid.h"
 
-void WaveTransformation(struct BeachGrid *grid, float wave_angle, float wave_period, float wave_height, float timestep_length);
-void GetAvailableSupply(struct BeachGrid *grid, int ref_pos, float ref_depth, float shelf_slope, float shoreface_slope, float min_depth);
+void WaveTransformation(struct BeachGrid *grid, double wave_angle, double wave_period, double wave_height, double timestep_length);
+void GetAvailableSupply(struct BeachGrid *grid, int ref_pos, double ref_depth, double shelf_slope, double shoreface_slope, double min_depth);
 void NetVolumeChange(struct BeachGrid *grid);
-int TransportSediment(struct BeachGrid *grid, int ref_pos, float ref_depth, float shelf_slope, float shoreface_slope, float min_depth);
+int TransportSediment(struct BeachGrid *grid, int ref_pos, double ref_depth, double shelf_slope, double shoreface_slope, double min_depth);
 int FixBeach(struct BeachGrid* grid);
 
 #if defined(__cplusplus)
