+++ server/C/cem/utils.c
@@ -25,9 +25,9 @@ void free2d(void **mem)
 /**
  * Generates a random number equally distributed between zero and one
  * PARAMETERS: none
- * RETURN: random float between zero and one
+ * RETURN: random double between zero and one
  */
-float RandZeroToOne(void)
+double RandZeroToOne(void)
 {
-	return (float)rand() / RAND_MAX;
+	return (double)rand() / RAND_MAX;
 }
